backend/python_services/search_query_service/requirements.txt
Flask>=3.0.0
chromadb>=0.5.0
python-dotenv>=1.0.1
google-generativeai>=0.5.2
gunicorn>=21.2.0
firebase-admin>=6.5.0